+++ core/jvm/src/test/scala/scalaz/zio/TestRuntime.scala
@@ -1,23 +1,16 @@
 package scalaz.zio
 
 import scala.concurrent.duration._
-import org.specs2.Specification
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.execute.{ AsResult, Failure, Result, Skipped }
 import org.specs2.matcher.Expectations
 import org.specs2.matcher.TerminationMatchers.terminate
 import org.specs2.specification.{ Around, AroundEach, AroundTimeout }
-import scalaz.zio.internal.PlatformLive
 
 abstract class TestRuntime(implicit ee: org.specs2.concurrent.ExecutionEnv)
-    extends Specification
-    with DefaultRuntime
+    extends BaseCrossPlatformSpec
     with AroundEach
     with AroundTimeout {
-
-  override val Platform = PlatformLive.makeDefault().withReportFailure(_ => ())
-
-  val DefaultTimeout = 60.seconds
 
   override final def around[R: AsResult](r: => R): Result =
     AsResult.safely(upTo(DefaultTimeout)(r)) match {
