+++ core/shared/src/test/scala/scalaz/zio/BaseCrossPlatformSpec.scala
@@ -17,7 +17,7 @@ import scalaz.zio.system.System
 import scala.concurrent.duration.{ Duration, _ }
 import scala.concurrent.{ ExecutionContext, Future }
 
-abstract class BaseCrossPlatformSpec extends Specification with DefaultRuntime with TestUtils {
+abstract class BaseCrossPlatformSpec extends Specification with DefaultRuntime {
 
   override val Platform = PlatformLive.makeDefault().withReportFailure(_ => ())
 
@@ -25,6 +25,19 @@ abstract class BaseCrossPlatformSpec extends Specification with DefaultRuntime w
 
   val DefaultTimeout: Duration = 60.seconds
   val timer                    = new Timer()
+
+  def withLatch[R, E, A](f: UIO[Unit] => ZIO[R, E, A]): ZIO[R, E, A] =
+    Promise.make[Nothing, Unit] >>= (latch => f(latch.succeed(()).unit) <* latch.await)
+
+  def withLatch[R, E, A](f: (UIO[Unit], UIO[Unit]) => ZIO[R, E, A]): ZIO[R, E, A] =
+    for {
+      ref   <- Ref.make(true)
+      latch <- Promise.make[Nothing, Unit]
+      a <- f(latch.succeed(()).unit, ZIO.uninterruptibleMask { restore =>
+            ref.set(false) *> restore(latch.await)
+          })
+      _ <- UIO.whenM(ref.get)(latch.await)
+    } yield a
 
   implicit def zioAsExecution[A: AsResult, R >: Clock with Console with System with Random, E]
     : AsExecution[ZIO[R, E, A]] =
